[GraphScheduler] Add a simple topological sort graph scheduler
