+++ lib/IR/TopologicalSortBasedScheduler.cpp
@@ -1,0 +1,36 @@
+/**
+ * Copyright (c) 2017-present, Facebook, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "GraphScheduler.h"
+
+#include "glow/Graph/Utils.h"
+
+#include <queue>
+
+using llvm::isa;
+
+namespace glow {
+void TopologicalSortBasedScheduler::schedule() {
+  GraphPostOrderVisitor gpov(G_);
+  auto PostOrderNodes = gpov.getPostOrder();
+  for (auto &N : PostOrderNodes) {
+    if (isa<Storage>(N)) {
+      continue;
+    }
+    scheduled_.push_back(N);
+  }
+}
+} // namespace glow
